+++ tests/testthat/test-mock-api.R
@@ -9,9 +9,10 @@ with_mock_api({
     obj <- request("api/object1/") %>%
       req_url_query(a = 1) %>%
       req_perform()
-    expect_json_equal(
+    expect_equal(
       resp_body_json(obj),
-      list(query = list(a = 1), mocked = "yes")
+      list(query = list(a = 1), mocked = "yes"),
+      list_as_map = TRUE
     )
   })
   test_that("GET with special characters", {